# Implements prefilter and configures TinyGo unit tests