--- a/examples/score-simple/main.go
+++ b/examples/score-simple/main.go
@@ -16,7 +16,11 @@
 
 package main
 
+// Override the default GC with a more performant one.
+// Note: this requires tinygo flags: -gc=custom -tags=custommalloc
 import (
+	_ "github.com/wasilibs/nottinygc"
+
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/score"
 )