--- a/guest/internal/imports/imports.go
+++ b/guest/internal/imports/imports.go
@@ -18,14 +18,16 @@
 
 package imports
 
+import "sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/mem"
+
 //go:wasmimport k8s.io/api nodeInfo/node
-func k8sApiNodeInfoNode(ptr uint32, limit bufLimit) (len uint32)
+func k8sApiNodeInfoNode(ptr uint32, limit mem.BufLimit) (len uint32)
 
 //go:wasmimport k8s.io/api nodeName
-func k8sApiNodeName(ptr uint32, limit bufLimit) (len uint32)
+func k8sApiNodeName(ptr uint32, limit mem.BufLimit) (len uint32)
 
 //go:wasmimport k8s.io/api pod
-func k8sApiPod(ptr uint32, limit bufLimit) (len uint32)
+func k8sApiPod(ptr uint32, limit mem.BufLimit) (len uint32)
 
-//go:wasmimport k8s.io/scheduler status_reason
-func k8sSchedulerStatusReason(ptr, size uint32)
+//go:wasmimport k8s.io/scheduler result.status_reason
+func k8sSchedulerResultStatusReason(ptr, size uint32)