--- a/guest/testdata/all-noop/main.go
+++ b/guest/testdata/all-noop/main.go
@@ -16,9 +16,14 @@
 
 package main
 
+// Override the default GC with a more performant one.
+// Note: this requires tinygo flags: -gc=custom -tags=custommalloc
 import (
+	_ "github.com/wasilibs/nottinygc"
+
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/filter"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/prefilter"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/score"
 )
 
@@ -27,9 +32,11 @@ func main() {
 	// shows the impact two things:
 	//  * implementing multiple interfaces
 	//  * overhead of constructing function parameters
+	prefilter.Plugin = api.PreFilterFunc(prefilterNoop)
 	filter.Plugin = api.FilterFunc(filterNoop)
 	score.Plugin = api.ScoreFunc(scoreNoop)
 }
 
-func filterNoop(api.Pod, api.NodeInfo) (status *api.Status)       { return }
-func scoreNoop(api.Pod, string) (score int32, status *api.Status) { return }
+func prefilterNoop(pod api.Pod) (nodeNames []string, status *api.Status) { return }
+func filterNoop(api.Pod, api.NodeInfo) (status *api.Status)              { return }
+func scoreNoop(api.Pod, string) (score int32, status *api.Status)        { return }