--- a/guest/testdata/go.mod
+++ b/guest/testdata/go.mod
@@ -1,4 +1,4 @@
-module sigs.k8s.io/kube-scheduler-wasm-extension/examples
+module sigs.k8s.io/kube-scheduler-wasm-extension/guest/testdata
 
 go 1.20
 
@@ -13,6 +13,6 @@ require (
 	sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto v0.0.0-00010101000000-000000000000 // indirect
 )
 
-replace sigs.k8s.io/kube-scheduler-wasm-extension/guest => ./../guest
+replace sigs.k8s.io/kube-scheduler-wasm-extension/guest => ./../
 
-replace sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto => ./../kubernetes/proto
+replace sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto => ./../../kubernetes/proto