… .github/workflows/commit.yaml
… examples/filter-simple/main.wasm
… examples/prefilter-simple/main.go
… examples/prefilter-simple/main.wasm
… examples/score-simple/main.wasm
… guest/RATIONALE.md
… guest/prefilter/cstring_test.go
… guest/testdata/all-noop/main.wasm
… guest/testdata/all/main.wasm
… guest/testdata/filter/main.wasm
… guest/testdata/go.sum
… guest/testdata/prefilter/main.wasm
… guest/testdata/score/main.wasm
… internal/e2e/plugin_perf_test.go
… scheduler/plugin/cstring_test.go
… scheduler/plugin/host.go
… scheduler/plugin/plugin.go
… scheduler/plugin/plugin_test.go
… scheduler/test/testdata.go
… scheduler/test/testdata/error/panic_on_filter.wasm
… scheduler/test/testdata/error/panic_on_prefilter.wasm
… scheduler/test/testdata/error/panic_on_prefilter.wat
… scheduler/test/testdata/error/panic_on_score.wasm
… scheduler/test/testdata/error/panic_on_start.wasm
… scheduler/test/testdata/error/panic_on_start.wat
… scheduler/test/testdata/test/all_noop.wasm
… scheduler/test/testdata/test/all_noop.wat
… scheduler/test/testdata/test/prefilter_from_global.wasm
… scheduler/test/testdata/test/prefilter_from_global.wat